+++ tests/testthat/test-calculate.R
@@ -613,3 +613,18 @@ test_that("calculate messages informatively with excessive null", {
     "point null hypothesis `sigma = 10` does not inform calculation"
   )
 })
+
+test_that("calculate can handle variables named x", {
+  expect_silent({
+    t_0 <- data.frame(x = 1:10) %>% 
+      specify(response = x) %>% 
+      hypothesise(null = "point", mu = 0) %>% 
+      calculate(stat = "t")
+  })
+  
+  expect_equal(
+    unname(t_0$stat),
+    5.74,
+    tolerance = .001
+  )
+})